+++ CHANGELOG.md
@@ -6,8 +6,10 @@
 - Do not put empty uid/gid corosync configuration options to an uidgid file
   when not specified in `pcs cluster uidgid add` command. Empty options cause
   corosync start failure. ([ghissue#772])
+- Do not allow fencing levels other than 1..9 ([RHEL-2977])
 
 [ghissue#772]: https://github.com/ClusterLabs/pcs/issues/772
+[RHEL-2977]: https://issues.redhat.com/browse/RHEL-2977
 
 
 ## [0.11.7] - 2024-01-11